+++ src/components/chart/Chart.jsx
@@ -7,37 +7,7 @@ import { GraphContext } from "../../context/graphContext/GraphContext";
 import { isNotNullOrUndefined } from "../../hooks/nullOrUndefined";
 import { NodeMenu } from "./Node/NodeMenu";
 
-export const downloadSvgAsImage = (svgRef, format = "png") => {
-  if (!svgRef.current) return;
-
-  const svgData = new XMLSerializer().serializeToString(svgRef.current);
-  const svgBlob = new Blob([svgData], {
-    type: "image/svg+xml;charset=utf-8",
-  });
-  const svgUrl = URL.createObjectURL(svgBlob);
-
-  const canvas = document.createElement("canvas");
-  canvas.width = svgRef.current.clientWidth;
-  canvas.height = svgRef.current.clientHeight;
-
-  const ctx = canvas.getContext("2d");
-  const img = new Image();
-  img.onload = () => {
-    ctx?.drawImage(img, 0, 0);
-    URL.revokeObjectURL(svgUrl);
-
-    const imageUrl = canvas.toDataURL(`image/${format}`);
-    const link = document.createElement("a");
-    link.href = imageUrl;
-    link.download = `graph.${format}`;
-    document.body.appendChild(link);
-    link.click();
-    document.body.removeChild(link);
-  };
-  img.src = svgUrl;
-};
-
-export const Chart = () => {
+export const Chart = ({ svgRef }) => {
   const {
     nodes,
     edges,
@@ -53,7 +23,6 @@ export const Chart = () => {
     isDrawEdgeMode,
   } = useContext(GraphContext);
 
-  const svgRef = useRef(null);
   const nodeMenuRef = useRef(null); // NodeMenu用のref
   const { width: svgWidth, height: svgHeight } = useSvgSize(svgRef);
 
@@ -159,7 +128,7 @@ export const Chart = () => {
         height="100%"
         tabIndex={0}
         onDoubleClick={handleDoubleClick}
-        style={{ clickEvent: "none", outline: "none" }}
+        style={{ clickEvent: "none", outline: "none", background: "white" }}
       >
         {edges.map((edge, i) => (
           <Edge
@@ -182,7 +151,6 @@ export const Chart = () => {
           />
         ))}
       </svg>
-
       <NodeMenu
         nodeMenuRef={nodeMenuRef}
         isOpen={sideMenuVisible}
